Fix Python 2 pathname type set by from_archive()
A trailing comma resulted in Entry.pathname being a tuple instead
of a string when using Entry.from_archive() in Python 2.

This partially reverts 501316595815 ("Fix tests for python 2 and 3"),
which accounted for Entry.pathname as a tuple, instead of addressing why
it had become one.

This also adds whitespace around the assignment operator for Python 2
and 3 to look less like a keyword argument assignment (which appears to
be where the comma originally came from).